+++ src/data/sparse_page_dmatrix.h
@@ -69,7 +69,7 @@ class SparsePageDMatrix : public DMatrix {
   float missing_;
   int nthreads_;
   std::string cache_prefix_;
-  size_t n_batches_ {0};
+  uint32_t n_batches_ {0};
   // sparse page is the source to other page types, we make a special member function.
   void InitializeSparsePage();
   // Non-virtual version that can be used in constructor
